requirements.txt
# FullFront Project Dependencies
# Model Response Generation
anthropic>=0.8.0
openai>=1.0.0
google-generativeai>=0.3.0
transformers>=4.35.0
pillow>=9.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.30.0

# HTML Rendering
playwright>=1.40.0

# Similarity Calculation
torch>=2.0.0
torchvision>=0.15.0
clip-openai>=1.0
opencv-python>=4.8.0
sentence-transformers>=2.2.2
bs4>=0.0.1
beautifulsoup4>=4.12.0
html-similarity>=0.3.3
lxml>=4.9.0
difflib>=0.1.0

# Data Processing
pyarrow>=14.0.0
fastparquet>=2023.10.0 